The shell now looks like it was made by a CS101 student!
The current framework of the shell is:
    1. Collect user input one character at a time and tokenize them on the fly.
    1a. If the user has not entered anything on the current line and the
        current is EOF, then the shell will exit. There is no other way of
        exiting the shell cleanly currently.
    2. When the user inputs a newline character that isn't enclosed in quotes
       then process the current list of tokens. Whitespace characters that are
       not in quotes will not be a part of any token.
    3. "Processing" the tokens is currently outputing a comma separated list
       of the tokens input.
    4. Clean up the previous command and wait for more input (goto 1).

I added a Log function that tracks where it was called from and can print
arbitrary printf formats to a log file. Currently the log file is
"my_shell.log" in the directory you're running the shell in. This will
likely become a command line argument in the future when I get far enough
to have a reason to doing arg stuff.